view extension
implementation of 6 images is missing
