--- Crypt Creeper/Crypt Creeper/HowToPlay/HowToPlayExtencion.swift
@@ -1,12 +1,0 @@
-//
-//  HowToPlayExtencion.swift
-//  Crypt Creeper
-//
-//  Created by Osasu sanchez on 25/2/23.
-//
-
-import Foundation
-
-extension HowToPlayView {
-    
-}
